e2e: rebalance sharded suites
